First pass at opcodes and operands
This is a first pass at the operands and opcodes needed by the jsonnet
spec. There is at least one optimization of this. When comp;iling a
`Call`, instead of pushing the function on to the stack first, we
would want an `Invoke` method for this. We will get there