Get proper Error reporting
With this commit, it refactors so that the `Compiler` uses the
`Parser` to get `TokenInfo` objects. Then it uses those `TokenInfo`
objects to track the `Span` of any errors.

Combing that `Span` with the `source_id` located in the `Scanner` this
now produces nice error messages like:
```
❌ Compilation failed:
Error: Invalid expression starting with Eof
   ╭─[ repl:1:5 ]
   │
 1 │ (1 +
   │     │
   │     ╰─ Invalid expression starting with Eof
───╯
```

and

```
❌ Compilation failed:
Error: Expected closing parenthesis, found Eof
   ╭─[ /home/scott/Projects/RapidJsonnet/end2end/complex_expression.jsonnet:1:17 ]
   │
 1 │ -3 * (4 - 1 + 2
   │                 │
   │                 ╰─ Expected closing parenthesis, found Eof
───╯
```