Add a debug compilation method
This method is for visulizating the compiled chunk. The goal of this
output is to inspect the generated code and confirm it "makes sense".

In the near future, there will be some debug added to track the
virtual machine state -- stack, and current program counter -- to
ensure it calculates the right value