Refactor of Value into Chunk and more types
Prior to this commit we had two `Value` definitions by mistake. The
VirtualMachine had one that wasn't really used and `Chunk` has one
type defed to f64. This changes that so we use the enum and start
packing in the basic values supported by Jsonnet -- well except
strings which are going to be challenging.