Don't always run GC for objects
Prior to this commit, every run through the byte code resulted in a
garbage collection pass over objects. This is because the
VirtualMachine didn't implement any threshold based configuration.

In this change, there is a threshold based configuration to avoid
doing garbage collection. All the Bazel targets don't use that ATM
because of the `stress_gc` feature but that will move to a seperate
end2end testing suite at a later date.