Indicate if virtual machine should do a garbage collection
Garbage collection only "makes sense" during execution. Therefore, we
should indicate when garbage collection should run but do not execute
it and defer it to logic in virtual_machine.

Move the properties hashmap to use a StringIndex instead of the
managed string directly so that we can do the garbage collection pass
and mark it when we execute