Garbage collection implementation
This takes a set of Values as roots and iterates over them. For any
object in that root list, it adds field names to the deque to get
marked.

After marking, iterate over each string and object, if they are not
marked then delete them. If they are, unmark them for the next sweep
phase.