Add Arrays creation and concat. Plus run all end2end tests
This adds arrays handling and adds a test target to run all end2end
tests. The arrays are garbage collected objects that hold other
Values. These can be concated together and output into a json form.

This change also adds a test target to run all the end2end
tests. The file names that should fail don't cause the test to
fail. and file names that should fail but don't cause the test to fail